Move PCBundleDoc to file
